fix some bugs like convert national cod format from int32 to int 64 (long)

+++ MyApplication/UpdateUserForm.cs
@@ -95,7 +95,7 @@ public partial class UpdateUserForm : BaseForm
             currentUser.FirstName = firstNameTextBox.Text;
             currentUser.LastName = lastNameTextBox.Text;
             currentUser.FullName = fullNameTextBox.Text;
-            currentUser.NationalCod = Convert.ToInt32(nationalCodeTextBox.Text);
+            currentUser.NationalCod = Convert.ToInt64(nationalCodeTextBox.Text);
             currentUser.Address = addressTextBox.Text;
 
             applicationDbContext.SaveChanges();
